Added Boss and bullet enemy

+++ Game_1/src/com/github/raweber/java17/BasicEnemy.java
@@ -13,7 +13,7 @@ public class BasicEnemy extends GameObject {
 	}
 
 	public Rectangle getBounds() {
-		return new Rectangle(x, y, 16, 16);
+		return new Rectangle((int)x, (int)y, 16, 16);
 	}
 
 	public void tick() {
@@ -25,11 +25,11 @@ public class BasicEnemy extends GameObject {
 		if (x <= 0 || x >= Game.WIDTH - 16)
 			velX *= -1;
 
-		handler.addObject(new Trail(x, y, ID.Trail, Color.red, 16, 16, 0.03f, handler));
+		handler.addObject(new Trail(x, y, ID.Trail, Color.red, 16, 16, 0.03, handler));
 	}
 
 	public void render(Graphics g) {
 		g.setColor(Color.red);
-		g.fillRect(x, y, 16, 16);
+		g.fillRect((int)x, (int)y, 16, 16);
 	}
 }
